requirements.txt
openai-whisper
numpy
sounddevice
scipy
pyttsx3
beautifulsoup4
wikipedia
requests
